added consistency using AuthGuard wrapping
