+++ apps/web/app/(auth)/sign-in/[[...sign-in]]/page.tsx
@@ -1,8 +1,8 @@
-import { SignIn } from "@clerk/nextjs"
+import { SignInView } from "@/modules/auth/ui/views/sign-in-view"
 
 const Page = () => {
     return (
-        <SignIn />
+        <SignInView />
     )
 }
