+++ apps/web/app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -1,8 +1,8 @@
-import { SignUp } from "@clerk/nextjs"
+import { SignUpView } from "@/modules/auth/ui/views/sign-up-view"
 
 const Page = () => {
     return (
-        <SignUp />
+        <SignUpView />
     )
 }
