+++ apps/web/app/(dashboard)/layout.tsx
@@ -1,0 +1,11 @@
+import { AuthGuard } from "@/modules/auth/ui/components/auth-guard"
+
+const Layout = ({ children }: { children: React.ReactNode }) => {
+    return (
+        <AuthGuard>
+            {children}
+        </AuthGuard>
+    )
+}
+
+export default Layout
