+++ apps/web/app/(dashboard)/page.tsx
@@ -1,0 +1,23 @@
+"use client"
+
+import { SignIn, SignInButton, UserButton } from "@clerk/nextjs"
+import { useQuery, useMutation } from "convex/react" 
+import { api } from "@workspace/backend/_generated/api"
+import { Button } from "@workspace/ui/components/button"
+
+export default function Page() {
+  
+  const users = useQuery(api.users.getMany)
+  const addUser = useMutation(api.users.add)
+
+  return (
+    <div className="flex items-center justify-center min-h-svh">
+      <p>apps/web</p>
+      <UserButton/>
+      <Button onClick={() => addUser()}>Add</Button>
+      <div className="max-w-sm w-full mx-auto">
+        {JSON.stringify(users, null, 2)}
+      </div>
+    </div>
+  )
+}
