--- apps/web/app/test/page.tsx
@@ -1,9 +1,0 @@
-const Page = () => {
-    return (
-        <div>
-            Test page!
-        </div>
-    )
-}
-
-export default Page
