+++ apps/web/modules/auth/ui/components/auth-guard.tsx
@@ -1,0 +1,27 @@
+"use client"
+
+import { Authenticated, Unauthenticated, AuthLoading } from "convex/react"
+import { AuthLayout } from "../layouts/auth_layout"
+import { SignInView } from "../views/sign-in-view"
+
+export const AuthGuard = ({ children }: { children: React.ReactNode }) => {
+    return (
+        <>
+        <AuthLoading>
+            <AuthLayout>
+                <p>Loading...</p>
+            </AuthLayout>
+        </AuthLoading>
+
+        <Authenticated>
+            {children}
+        </Authenticated>
+
+        <Unauthenticated>
+            <AuthLayout>
+                <SignInView/>
+            </AuthLayout>
+        </Unauthenticated>
+        </>
+    )
+}
