+++ apps/web/modules/auth/ui/layouts/auth_layout.tsx
@@ -1,0 +1,7 @@
+export const AuthLayout = ({ children }: { children: React.ReactNode }) => {
+    return (
+        <div className="min-h-screen min-w-screen h-full flex flex-col items-center justify-center">
+            {children}
+        </div>
+    )
+}
