+++ apps/web/modules/auth/ui/views/sign-in-view.tsx
@@ -1,0 +1,7 @@
+import { SignIn } from "@clerk/nextjs"
+
+export const SignInView = () => {
+    return (
+        <SignIn routing="hash"/>
+    )
+}
