+++ apps/web/modules/auth/ui/views/sign-up-view.tsx
@@ -1,0 +1,7 @@
+import { SignUp } from "@clerk/nextjs"
+
+export const SignUpView = () => {
+    return (
+        <SignUp routing="hash"/>
+    )
+}
